backend/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-cors-headers==2.4.0
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
drf-yasg==1.9.1
future==0.17.1
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2019.1
requests==2.19.1
ruamel.yaml==0.15.97
six==1.12.0
uritemplate==3.0.0
urllib3==1.23
